Refactor middleware into gql package
