+++ server/main.go
@@ -10,7 +10,6 @@ import (
 	"github.com/maxbrain0/react-go-graphql/server/data"
 	"github.com/maxbrain0/react-go-graphql/server/gql"
 	"github.com/maxbrain0/react-go-graphql/server/logger"
-	"github.com/maxbrain0/react-go-graphql/server/middleware"
 	"github.com/sirupsen/logrus"
 )
 
@@ -38,7 +37,7 @@ func main() {
 		Playground: true,
 	})
 
-	http.Handle("/graphql", middleware.HTTPMiddleware(h))
+	http.Handle("/graphql", gql.HTTPMiddleware(h))
 
 	ctxLogger.WithFields(logrus.Fields{
 		"port": port,
